+++ src/builders/AppBuilder.ts
@@ -1,5 +1,6 @@
 import {defineValue} from "@bejibun/utils";
 import path from "path";
+import config from "@/config/app";
 
 export default class AppBuilder {
     protected cwd: string;
@@ -8,6 +9,9 @@ export default class AppBuilder {
         this.cwd = process.cwd();
     }
 
+    /**
+     * Path Manager
+     */
     public appPath(_path?: string): string {
         return path.resolve(this.cwd, `app/${defineValue(_path, "")}`);
     }
@@ -46,5 +50,12 @@ export default class AppBuilder {
 
     public storagePath(_path?: string): string {
         return path.resolve(this.cwd, `storage/${defineValue(_path, "")}`);
+    }
+
+    /**
+     * Utils
+     */
+    public async isMaintenanceMode(): Promise<boolean> {
+        return Bun.file(config.maintenance.file).exists();
     }
 }