+++ src/facades/App.ts
@@ -1,6 +1,9 @@
 import AppBuilder from "@/builders/AppBuilder";
 
 export default class App {
+    /**
+     * Path Manager
+     */
     public static appPath(path?: string): string {
         return new AppBuilder().appPath(path);
     }
@@ -39,5 +42,12 @@ export default class App {
 
     public static storagePath(path?: string): string {
         return new AppBuilder().storagePath(path);
+    }
+
+    /**
+     * Utils
+     */
+    public static async isMaintenanceMode(): Promise<boolean> {
+        return await new AppBuilder().isMaintenanceMode();
     }
 }